+++ concourse-driver-java/src/main/java/com/cinchapi/concourse/CachedConnectionPool.java
@@ -16,6 +16,7 @@
 package com.cinchapi.concourse;
 
 import java.util.Queue;
+import java.util.function.Supplier;
 
 import com.cinchapi.concourse.util.ConcurrentLoadingQueue;
 
@@ -54,8 +55,18 @@ class CachedConnectionPool extends ConnectionPool {
      */
     protected CachedConnectionPool(String host, int port, String username,
             String password, String environment, int poolSize) {
-        super(() -> Concourse.connect(host, port, username, password,
+        this(() -> Concourse.connect(host, port, username, password,
                 environment), poolSize);
+    }
+
+    /**
+     * Construct a new instance.
+     * 
+     * @param supplier
+     * @param poolSize
+     */
+    protected CachedConnectionPool(Supplier<Concourse> supplier, int poolSize) {
+        super(supplier, poolSize);
     }
 
     @Override